+++ src/merc.h
@@ -2105,6 +2105,7 @@ bool	check_ban	args( ( char *site, int type) );
 
 
 /* comm.c */
+int     isutf8  args((char c));
 void	show_string	args( ( struct descriptor_data *d, char *input) );
 void	close_socket	args( ( DESCRIPTOR_DATA *dclose ) );
 void	write_to_buffer	args( ( DESCRIPTOR_DATA *d, const char *txt,